git-svn-id: https://svn.opendtect.org/od/trunk@32069 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiCoin/uistereodlg.cc
@@ -38,8 +38,8 @@ void uiStereoDlg::doFinalise( CallBacker* )
     int maxval = (int)(offset + 1000); maxval /= 100; maxval *= 100;
     if ( minval < 10 ) minval = 10;
 
-    sliderfld->sldr()->setMinValue( minval );
-    sliderfld->sldr()->setMaxValue( maxval );
+    sliderfld->sldr()->setMinValue( (float) minval );
+    sliderfld->sldr()->setMaxValue( (float) maxval );
     sliderfld->sldr()->setValue( offset );
 }
